Move Connection2 to Redis.cc/hh
